// src/chai.d.ts
export interface Chai {
  expect: ExpectStatic;
  should(): Should;
  use(fn: (chai: Chai, util: ChaiUtil) => void): Chai;
  assert: AssertStatic;
  config: Config;
  Assertion: AssertionConstructor;
  AssertionError: typeof AssertionError;
  version: string;
}

export interface ExpectStatic extends AssertionStatic {
  fail(actual?: any, expected?: any, message?: string, operator?: Operator): void;
}

export interface AssertStatic extends Assert {
}

export interface AssertionStatic {
  (target: any, message?: string): Assertion;
}

export type Operator = "==" | "===" | ">" | ">=" | "<" | "<=" | "!=" | "!==";

export type OperatorComparable = boolean | null | number | string | undefined | Date;

interface ShouldAssertion {
  equal(value1: any, value2: any, message?: string): void;
  Throw: ShouldThrow;
  throw: ShouldThrow;
  exist(value: any, message?: string): void;
}

interface Should extends ShouldAssertion {
  not: ShouldAssertion;
  fail(actual: any, expected: any, message?: string, operator?: Operator): void;
}

interface ShouldThrow {
  (actual: Function, expected?: string | RegExp, message?: string): void;
  (actual: Function, constructor: Error | Function, expected?: string | RegExp, message?: string): void;
}

export interface AssertionConstructor {
  new(value: unknown): Assertion;
}

export interface Assertion extends LanguageChains, NumericComparison, TypeComparison {
  not: Assertion;
  deep: Deep;
  ordered: Ordered;
  nested: Nested;
  any: KeyFilter;
  all: KeyFilter;
  a: TypeComparison;
  an: TypeComparison;
  include: Include;
  includes: Include;
  contain: Include;
  contains: Include;
  ok: Assertion;
  true: Assertion;
  false: Assertion;
  null: Assertion;
  undefined: Assertion;
  NaN: Assertion;
  exist: Assertion;
  empty: Assertion;
  arguments: Assertion;
  Arguments: Assertion;
  equal: Equal;
  equals: Equal;
  eq: Equal;
  eql: Equal;
  eqls: Equal;
  property: Property;
  ownProperty: OwnProperty;
  haveOwnProperty: OwnProperty;
  ownPropertyDescriptor: OwnPropertyDescriptor;
  haveOwnPropertyDescriptor: OwnPropertyDescriptor;
  length: Length;
  lengthOf: Length;
  match: Match;
  matches: Match;
  string(string: string, message?: string): Assertion;
  keys: Keys;
  key(string: string): Assertion;
  throw: Throw;
  throws: Throw;
  Throw: Throw;
  respondTo: RespondTo;
  respondsTo: RespondTo;
  itself: Assertion;
  satisfy: Satisfy;
  satisfies: Satisfy;
  closeTo: CloseTo;
  approximately: CloseTo;
  members: Members;
  increase: PropertyChange;
  increases: PropertyChange;
  decrease: PropertyChange;
  decreases: PropertyChange;
  change: PropertyChange;
  changes: PropertyChange;
  extensible: Assertion;
  sealed: Assertion;
  frozen: Assertion;
  oneOf(list: ReadonlyArray<any>, message?: string): Assertion;
}

interface LanguageChains {
  to: Assertion;
  be: Assertion;
  been: Assertion;
  is: Assertion;
  that: Assertion;
  which: Assertion;
  and: Assertion;
  has: Assertion;
  have: Assertion;
  with: Assertion;
  at: Assertion;
  of: Assertion;
  same: Assertion;
  but: Assertion;
  does: Assertion;
}

interface NumericComparison {
  above: NumberComparer;
  gt: NumberComparer;
  greaterThan: NumberComparer;
  least: NumberComparer;
  gte: NumberComparer;
  below: NumberComparer;
  lt: NumberComparer;
  lessThan: NumberComparer;
  most: NumberComparer;
  lte: NumberComparer;
  within(start: number, finish: number, message?: string): Assertion;
  within(start: Date, finish: Date, message?: string): Assertion;
}

interface NumberComparer {
  (value: number | Date, message?: string): Assertion;
}

interface TypeComparison {
  (type: string, message?: string): Assertion;
  instanceof: InstanceOf;
  instanceOf: InstanceOf;
}

interface InstanceOf {
  (constructor: Object, message?: string): Assertion;
}

interface CloseTo {
  (expected: number, delta: number, message?: string): Assertion;
}

interface Nested {
  include: Include;
  property: Property;
  members: Members;
}

interface Deep {
  equal: Equal;
  equals: Equal;
  eq: Equal;
  include: Include;
  property: Property;
  members: Members;
  ordered: Ordered;
  nested: Nested;
}

interface Ordered {
  members: Members;
}

interface KeyFilter {
  keys: Keys;
  members: Members;
}

interface Equal {
  (value: any, message?: string): Assertion;
}

interface Property {
  (name: string, value?: any, message?: string): Assertion;
}

interface OwnProperty {
  (name: string, message?: string): Assertion;
}

interface OwnPropertyDescriptor {
  (name: string, descriptor: PropertyDescriptor, message?: string): Assertion;
  (name: string, message?: string): Assertion;
}

interface Length extends LanguageChains, NumericComparison {
  (length: number, message?: string): Assertion;
}

interface Include {
  (value: Object | string | number, message?: string): Assertion;
  keys: Keys;
  deep: Deep;
  ordered: Ordered;
  members: Members;
  any: KeyFilter;
  all: KeyFilter;
}

interface Match {
  (regexp: RegExp, message?: string): Assertion;
}

interface Keys {
  (...keys: string[]): Assertion;
  (keys: ReadonlyArray<any> | Object): Assertion;
}

interface Throw {
  (expected?: string | RegExp, message?: string): Assertion;
  (constructor: Error | Function, expected?: string | RegExp, message?: string): Assertion;
}

interface RespondTo {
  (method: string, message?: string): Assertion;
}

interface Satisfy {
  (matcher: Function, message?: string): Assertion;
}

interface Members {
  (set: ReadonlyArray<any>, message?: string): Assertion;
}

interface PropertyChange {
  (object: Object, property: string, message?: string): Assertion;
}
// ...
export interface Config {
  /**
   * Default: false
   */
  includeStack: boolean;

  /**
   * Default: true
   */
  showDiff: boolean;

  /**
   * Default: 40
   */
  truncateThreshold: number;
}

export class AssertionError {
  constructor(message: string, _props?: any, ssf?: Function);
  name: string;
  message: string;
  showDiff: boolean;
  stack: string;
}

interface Object {
  should: Assertion;
}

export interface ChaiUtil {
  /*!
  * test utility
  */
  test(...args: unknown[]): unknown;

  /*!
  * type utility
  */
  type(...args: unknown[]): unknown;

  /*!
  * expectTypes utility
  */  expectTypes(...args: unknown[]): unknown;

  /*!
  * message utility
  */
  getMessage(...args: unknown[]): unknown;

  /*!
  * actual utility
  */
  getActual(...args: unknown[]): unknown;

  /*!
  * Inspect util
  */
  inspect(...args: unknown[]): unknown;

  /*!
  * Object Display util
  */
  objDisplay(...args: unknown[]): unknown;

  /*!
  * Flag utility
  */
  flag(...args: unknown[]): unknown;

  /*!
  * Flag transferring utility
  */
  transferFlags(...args: unknown[]): unknown;

  /*!
  * Deep equal utility
  */
  eql(...args: unknown[]): unknown;

  /*!
  * Deep path info
  */
  getPathInfo(...args: unknown[]): unknown;

  /*!
  * Check if a property exists
  */
  hasProperty(...args: unknown[]): unknown;

  /*!
  * Function name
  */
  getName(...args: unknown[]): unknown;

  /*!
  * add Property
  */
  addProperty(...args: unknown[]): unknown;

  /*!
  * add Method
  */
  addMethod(...args: unknown[]): unknown;

  /*!
  * overwrite Property
  */
  overwriteProperty(...args: unknown[]): unknown;

  /*!
  * overwrite Method
  */
  overwriteMethod(...args: unknown[]): unknown;

  /*!
  * Add a chainable method
  */
  addChainableMethod(...args: unknown[]): unknown;

  /*!
  * Overwrite chainable method
  */
  overwriteChainableMethod(...args: unknown[]): unknown;

  /*!
  * Compare by inspect method
  */
  compareByInspect(...args: unknown[]): unknown;

  /*!
  * Get own enumerable property symbols method
  */
  getOwnEnumerablePropertySymbols(...args: unknown[]): unknown;

  /*!
  * Get own enumerable properties method
  */
  getOwnEnumerableProperties(...args: unknown[]): unknown;

  /*!
  * Checks error against a given set of criteria
  */
  checkError(...args: unknown[]): unknown;

  /*!
  * Proxify util
  */
  proxify(...args: unknown[]): unknown;

  /*!
  * addLengthGuard util
  */
  addLengthGuard(...args: unknown[]): unknown;

  /*!
  * isProxyEnabled helper
  */
  isProxyEnabled(...args: unknown[]): unknown;

  /*!
  * isNaN method
  */
  isNaN(...args: unknown[]): unknown;
}